# go-backend/docs/swagger.yaml
basePath: /
definitions:
...
  apidoc.CatalogCandidate:
    properties:
      data:
        additionalProperties: {}
        type: object
      equipment:
        type: string
      id:
        type: string
      match_type:
        type: string
      name:
        type: string
    type: object
  apidoc.CatalogExercise:
    properties:
      data:
        additionalProperties: {}
        type: object
      equipment:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  apidoc.CatalogSearch:
    properties:
      candidates:
        items:
          $ref: '#/definitions/apidoc.CatalogCandidate'
        type: array
      exercise_catalog_id:
        type: string
      match_type:
        type: string
      query:
        type: string
    type: object
...
  apidoc.CreatedID:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
  apidoc.Error:
    properties:
      error:
        type: string
    type: object
  apidoc.Exercise:
    properties:
      catalog_exercise:
        $ref: '#/definitions/apidoc.CatalogExercise'
      exercise_catalog_id:
        type: string
      id:
        type: integer
      name:
        type: string
      sets:
        items:
          $ref: '#/definitions/apidoc.ExerciseSet'
        type: array
    type: object
  apidoc.ExerciseBatchInput:
    properties:
      exercises:
        items:
          $ref: '#/definitions/apidoc.ExerciseInput'
        type: array
    type: object
  apidoc.ExerciseInput:
    properties:
      exercise_catalog_id:
        type: string
      name:
        type: string
      remember_alias:
        type: boolean
      sets:
        items:
          $ref: '#/definitions/apidoc.ExerciseSetInput'
        type: array
    type: object
  apidoc.ExerciseSet:
    properties:
      id:
        type: integer
      reps:
        type: integer
      set_number:
        type: integer
      weight:
        type: number
    type: object
  apidoc.ExerciseSetInput:
    properties:
      reps:
        type: integer
      weight:
        type: number
    type: object
  apidoc.FitnessOverview:
    properties:
      calendar_workouts:
        additionalProperties:
          $ref: '#/definitions/apidoc.FitnessWorkoutGroup'
        type: object
      cards:
        items:
          $ref: '#/definitions/apidoc.FitnessStatCard'
        type: array
      recent_workouts:
        items:
          $ref: '#/definitions/apidoc.FitnessWorkoutGroup'
        type: array
      today:
        type: string
    type: object
  apidoc.FitnessOverviewWorkout:
    properties:
      id:
        type: integer
      time_label:
        type: string
    type: object
  apidoc.FitnessStatCard:
    properties:
      id:
        type: string
      label:
        type: string
      subtitle:
        type: string
      tooltip:
        type: string
      value:
        type: string
    type: object
  apidoc.FitnessWorkoutDetail:
    properties:
      calories_burned:
        type: number
      created_at:
        type: string
      duration_seconds:
        type: number
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
    type: object
  apidoc.FitnessWorkoutGroup:
    properties:
      date:
        type: string
      date_label:
        type: string
      workouts:
        items:
          $ref: '#/definitions/apidoc.FitnessOverviewWorkout'
        type: array
    type: object
...
  apidoc.GymVisit:
    properties:
      created_at:
        type: string
      id:
        type: integer
    type: object
  apidoc.GymVisitDateInput:
    properties:
      date:
        example: "2026-10-01"
        type: string
    type: object
...
  apidoc.Visited:
    properties:
      id:
        type: integer
      visited:
        type: boolean
    type: object
...
info:
  contact: {}
  description: REST API for the Life Backend application. Session endpoints accept
    a Bearer token in Authorization, X-Session-Token, or the life_session cookie;
    login sets the cookie.
  title: Life Backend API
  version: "1.0"
paths:
...
  /api/exercise-catalog:
    get:
      description: Only unique exact names and confirmed personal aliases resolve
        automatically. Other candidates require selection.
      parameters:
      - description: Exercise name (1–100 characters)
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidoc.CatalogSearch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: Search exercise catalogue and personal aliases
      tags:
      - Gym
  /api/fitness-workouts/{id}:
    get:
      parameters:
      - description: Positive id
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidoc.FitnessWorkoutDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: Get workout details with linked fitness metrics
      tags:
      - fitness
  /api/fitness/overview:
    get:
      description: Uses Asia/Kolkata dates and year-to-date totals; excludes future
        entries. Weekly streaks use Monday-Sunday weeks. Timing and time preference
        use imported traditional strength sessions on recorded gym days. Cards disclose
        missing timing and weights.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidoc.FitnessOverview'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: Get render-ready gym statistics, calendar visits, and three recent
        workouts
      tags:
      - fitness
  /api/gym-visits:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apidoc.GymVisit'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: List gym visits
      tags:
      - gym
    post:
      consumes:
      - application/json
      parameters:
      - description: Workout date
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidoc.GymVisitDateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidoc.CreatedID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: Record a workout visit on an India calendar date
      tags:
      - gym
  /api/gym-visits/{id}:
    delete:
      parameters:
      - description: Positive id
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: Delete gym visit
      tags:
      - gym
  /api/gym-visits/{id}/exercises:
    get:
      parameters:
      - description: Positive id
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apidoc.Exercise'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: List visit exercises
      tags:
      - gym
    post:
      consumes:
      - application/json
      parameters:
      - description: Positive id
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidoc.ExerciseInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidoc.Exercise'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: Add visit exercise
      tags:
      - gym
  /api/gym-visits/{id}/exercises/{exerciseID}:
    delete:
      parameters:
      - description: Workout id
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Saved exercise id
        in: path
        minimum: 1
        name: exerciseID
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: Delete a saved exercise and its sets from a workout
      tags:
      - gym
  /api/gym-visits/{id}/exercises/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Positive id
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidoc.ExerciseBatchInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/apidoc.Exercise'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: Add multiple visit exercises
      tags:
      - gym
...
  /api/workout/today:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidoc.Visited'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: Get today's workout visit
      tags:
      - gym
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidoc.CreatedID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.Error'
      security:
      - SessionBearer: []
      summary: Record today's workout visit
      tags:
      - gym
...
schemes:
- http
- https
securityDefinitions:
  SessionBearer:
    description: Bearer session token. X-Session-Token and life_session cookie also
      work.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"